+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Change Log
 
 [upcoming release] - 2025-..-..
 -------------------------------
+- [FIXED] cim2pp: extracting the additional info about referencePrios, modes and targetValues for ext_grid / gen / sgen
 - [CHANGED] create.py: refactored to create module with smaller files. Backwards compatible, importing does not require change!
 - [FIXED] cim2pp: update docu with correct dependancy install
 - [FIXED] cim2pp: fixed using the trafo characteristics for the correct tap changer at 2w trafos